svn merge -r7515:HEAD http://svn.eol.ucar.edu/svn/raf/branches/vardb-20141208 .
Merge the latest vardb refactor branch onto the trunk.  The API is a little
different and so clients will need to be updated.  In particular, VDBVar
pointers are now owned by VDBFile and should not be deleted.  There are
also tests and a start at doxygen documentation.  See the docs index page
(or vardb.hh) for details of the changes.  Existing XML files should not
need changes.


git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk/vardb@7554 640d5228-2204-0410-8d71-8f46fa6850b4