fix inconsistency in include prefix for vardb headers
Since vardb headers will be installed and included with the 'raf/' prefix,
use that prefix everywhere in the source library also.  Otherwise some
source will find the headers from the installed location, causing scons to
install them first, rather than finding the headers in the source
directory.  Also, switch SConscript to tool_vardb.py so vardb can be
required from anywhere in the source tree, and if it is not found in the
source tree, then the parent source directory can switch to using the
installed vardb.